requirements.txt
mariadb==1.0.11
Mastodon.py==1.8.1
requests==2.28.1